Añadidos resto de logos y modificado tipo del dinero principal

+++ src/main/java/com/example/idleback/Dto/lenguaje_partida/Lenguaje_partidaDTO.java
@@ -10,7 +10,7 @@ public class Lenguaje_partidaDTO {
 
     private Long partidaId;
 
-    private Float partidaDinero;
+    private Double partidaDinero;
 
     private String lenguajeNombre;
